Make reducer directory and separate file
